use NLopt in fitvis
I have found that this is much more robust than the self-written
optimization routine I was using prior to this change. Unfortunately the
need to enforce an equality constraint greatly restricted the number of
algorithms from NLopt that were suitable for my use. I am therefore
using a gradient-free method that doesn't converge as quickly in the
best-case scenario, but performs a lot better in the worst-case
scenario.

See issue #56 for a description of some of the issues I ran into with
this.